[FIX] account_accountant: Fix computed field on bank reco widget line
- Create a statement line
- Create an invoice
- Match the statement line with the invoice in the widget without saving
=> The name of the invoice is displayed
- Validate
=> The name has changed

Explanation:
After the validation, the bank reco widget line is no longer a 'new_aml' but an 'aml' meaning its 'source_aml_id' that was an invoice is now the bank statement line journal entry itself.
So 'source_aml_move_id' must not be a related field but a computed field looking to the reconciliation when the bank reco widget line is 'aml'.

Also, remove the stats button that are redundant regarding the links set on the accounting items.

Task: 2942314
Part-of: odoo/enterprise#30490