[IMP] spreadsheet_edition: can use global filter without datasource
We now have some use cases that could benefit from the global filters,
for exemple ODOO.CREDIT function which can take the value from an
ODOO.FILTER.VALUE function.
This could work without a pivot, list or graph in the spreadsheet.
Therefore, we enable the use of a global filter without a data source.

Task-id 2954856

closes odoo/enterprise#30492

Signed-off-by: Rémi Rahir (rar) <rar@odoo.com>