[IMP] planning: change pop-up when we send shifts employees without email
In this commit we change the pop-up that is raised when we click on the
'Publish & Send' button of the 'Publish' dialog pop-up and we have
employees without an email address on record.
Previously, this would raise a different pop-up for each employee without
email in order for the email to be specified. Now, a single pop-up with
a list of employees and an input for their email is raised.

This should make filling multiple missing email addresses quicker and more
straightforward.

task-2952021

closes odoo/enterprise#30583

Signed-off-by: Laurent Stukkens (ltu) <ltu@odoo.com>