[FIX] social_marketing: make the "Add a stream" modal responsive
Currently, in social_marketing, when we try to add a stream on small devices,
the social platform cards are shown in single line and get out of the screen.

This commit fixes the issue by 'wrapping' the cards in next line if there is
not enough space. Also, as a side improvement, the width is now fixed at 90px,
so that when wrapped, the cards look aligned properly (having same width).
Also, Change the 'o_social_media_cards' class property to adapt the tour
according to the changes (to display the bubble in the proper position).

task-2893410

closes odoo/enterprise#31211

X-original-commit: d38cba20c160e3c8d0c68b70ad59b0ddf3647f3b
Signed-off-by: Thibault Delavallee (tde) <tde@openerp.com>