[FIX] appointment_crm: assign correct company to lead
Steps to recreate the issue:

- Log in as Mitchell
- Create a new company, let's call it c2
- c2 should now be an allowed company of Mitchell, but not their
preferred one.
- go to Settings > Manage Companies, drag c2 to the top of the list
- Go to Calendar > Online appointments
- Activate 'create opportunity'
- Via Website, create a new appointment as public user (log off
Mitchell), complete form (name, email, phone) and confirm appointment
- Sign back in as Mitchell
- Go to calendar: appointment is created with new contact (assignee)

Issue:

In CRM, the generated lead does not appear in Mitchell's preferred
company, but in c2. This happens because the lead's company_id
is chosen as env.company, which is set to the company with the
lowest sequence. Overall, this can confusing to the employee.

This commit adds the employee company as extra context to the lead creation.

opw-2951087

closes odoo/enterprise#34547

X-original-commit: fef7613705401783a425a52db2263cbdee5f5756
Signed-off-by: Nicolas Lempereur (nle) <nle@odoo.com>
Signed-off-by: Stefan-Calin Crainiciuc (stcc) <stcc@odoo.com>