[FIX] documents: disable facet quick create in actions list view
Using the quick create feature on facets in the workflow actions list view, it was possible to create facets not linked to any folder, bypassing the usual restriction in other views.
This generated bugs when duplicating folders.

This commit fixes the issue by disabling the quick create for facets in the action list view.

Task-2953966

closes odoo/enterprise#30529

X-original-commit: d2736ecdf9b67c9d986c4a397da62e7dc61146aa
Signed-off-by: William Braeckman (wbr) <wbr@odoo.com>
Signed-off-by: Larcin Vincent (vila) <vila@odoo.com>