[FIX] quality_mrp_workorder_iot: add missing dependencies
Not all iot extensions of the quality module or its dependents currently
include the iot counterparts of the dependencies of their parent modules.

For example: the quality_mrp_workorder_iot module only includes the quality_iot
and quality_mrp_workorder modules as its dependencies. Functionally though, it
also depends on the iot extension of the quality_control and mrp_workorder
modules. When these missing dependencies are not installed, this can result in
missing models or fields, causing a failure to load the registry.

Note: because these modules are generally automatically installed the bug does
not often manifest itself, however it can be triggered by manual install or
uninstall operations.

closes odoo/enterprise#32796

X-original-commit: 45b52ec7d7f8a7682b9bcc60cebc320640518f46
Signed-off-by: William Henrotin (whe) <whe@odoo.com>
Signed-off-by: De Caluwé Tom (tdc) <tdc@odoo.com>