[IMP] stock_barcode: remove paging by location
== No more page by locations ==
From this commit, there is only one page for the Barcode App. So, no
more pages aren't created by source/destination locations.
This change applies the following:
- The source and destination are displayed directly on the lines;
- As all the lines are now shown on the same page, the tracked products'
  lines are grouped also by source/destination locations.

== Locations displayed on the lines ==
Like said above, the source and destination are displayed on the lines.
That said, they won't be displayed if they are equal to the picking type
default source/destination location.
For the pickings, if the location is a sublocation of the picking's
location (e.g.: the picking's location is WH/Stock and the line's
location is WH/Stock/Shelf 2), only the sub part will be displayed (for
the previous example, it will display ".../Shelf 2").
Also, the location will be write in red if it's not an internal location
and the real location's name will write in bold (e.g.:
.../Shelf 2/*Small Refregirator*).

== Lines' order ==
The barcode line are sorted once at the opening and then they keep their
order even if their locations are changed or if they are completed.
That said, the order is recomputed when the application fetches the
record's data (for example, when a new record is created through the
form view).

== Scan a source location ==
When scanning a source location, the application will search for a line
in this location. If one is found, it will scroll to this line.

This commit also fixes or improves some behavior following previous
commits 868be02626e971ab528b720cd42218c59e288e39 and b8f6b5c417bc29ac10ffadf529f039593dc54e4a

Besides thoses changes, multiple tests were modified to match the one
page. Regarding this point, two tests were removed:
- The test `test_internal_picking_from_scratch_1` was removed as what it
  checks was not relevant anymore and the checks for the `write` count
  was moved in `test_internal_picking_from_scratch_2` as this part of
  the test is still relevant. Since this test is no more the second
  one, renames it `test_internal_picking_from_scratch`;
- The test `test_receipt_reserved_2` was removed for the same reason.

task-2960938
fp-request

Part-of: odoo/enterprise#30979