[FIX] knowledge: avoid archived article to be available for shared users.
The method _fetch_article was using browse instead of search. That means that
archived articles could be retrieved when accessing the route of the article.

As we don't make any difference between Forbidden and Not Found error, this
commit replaces the browse by search method that will automatically check access
rules but also filter out archived articles.

Task-2950423

Part-of: odoo/enterprise#30681