[FIX] web_enterprise: use ActionSwiper on mobile only
This commit fixes the settings form view. Since its
Owl conversion, the ActionSwiper component is used
in the template. Before this commit, it was used even
when we don't want the settings to allow the swipe.

It now only activate the swipe behavior on smaller
screens. A test have been added to make sure settings
are not swipable when settings are displayed on
a large screen.

The fix is only a temporary workaround as the component
crashed during its Owl lifecycle. The props isActive is
not intended to stay available as soon as the issue is
fixed. See issue at odoo/owl#1246

closes odoo/enterprise#30574

Signed-off-by: Aaron Bohy (aab) <aab@odoo.com>