[FIX] knowledge: fix command palette search rule
Purpose:
========
The article search rule in the command palette previously did not make
distinctions between articles whose name match the search query and
articles whose root's name match the search query. Therefore, the root
article matching the query was not always on the top since its children
also matched the query through their root.

Now, one first search for articles whose name match the query, and
afterwards search for articles whose name does not match the query but
whose root's name match the query.
The order of articles shown in the command palette is thus now:
- name = query & is_user_favorite - by Favorite sequence
- name = query & Favorite count
- root.name = query & is_user_favorite - by Favorite sequence
- root.name = query & Favorite count

Tests have been adapted so that articles whose name match the query are
always expected before articles whose root.name match the query.

The number of articles shown in the command palette has been increased
to 40 instead of 10, to ease the search.

The matching part of the root name is now also highlighted in the
command palette, so that one gets why the article appears there. It is
however not bold anymore as we want to keep it light compared to the
article name.

Task-2922605

closes odoo/enterprise#30646

X-original-commit: 832626b77f3391dd7ae9cd388723200cb8508777
Signed-off-by: Thibault Delavallee (tde) <tde@openerp.com>