[FIX] knowledge: fix article title selector for default title
When an article does not have any title and the user clicks on the text
input containing the title of the document, the script will automatically
fill the text input with the first h1 of the document.

The selector allowing the script to retrieve the first h1 is not correct
and can select several titles at once. When it occurs, the script fills
the text input with a concatenation of all h1 titles which does not
correspond to the desired behavior.

With this commit, we will update the selector to select the first h1 of
the document. We will also slightly refactor the function to make sure
that the model is updated before updating the view.

task-2957260

closes odoo/enterprise#30612

X-original-commit: 651ee66f22b5ad8ce5a73ae383db6705ad193e9d
Signed-off-by: David Beguin (dbe) <dbe@odoo.com>
Signed-off-by: Thibault Delavallee (tde) <tde@openerp.com>