[IMP] marketing_automation_*: new dynamic placeholder
Introduce a new Dynamic Placeholder Generator
in mailing and SMS template creation.

Using a specific character (#) in sms body will trigger
the opening of the field selector popup, which can be use to select
the wanted field and enter a default value. on validation the dynamic
placeholder will be inserted in the corresponding textarea.

task-2154624

closes odoo/enterprise#30333

Related: odoo/odoo#71878
Related: odoo/upgrade#3766
Signed-off-by: David Monjoie (dmo) <dmo@odoo.com>