[FIX] test_l10n_be_hr_payroll_account: avoid failing uninstall
Reinstalling the test_l10n_be_hr_payroll_account fails when calling the post
init hook. The failure happens because of a constraint violation caused by
data left behind after the first uninstall. As a result the registry cannot be
loaded.

The problem is resolved by assigning xml ids to the population data, so that
the data is properly removed when uninstalling the module.

closes odoo/enterprise#30668

Signed-off-by: Yannick Tivisse (yti) <yti@odoo.com>