[FIX] project_account_budget: review budget calculation
Before this commit, the remaining budget was wrongly calculation because in
some case the allocated budget could be the one estimated and so the amount
could be less than 0, and so if the budget spent does not spent more than
the amount set on the allocated budget then the value in the remaining
budget should not the allocated budget minus the budget spent, otherwise
the remaining budget will show the budget spent was
(budget allocated + budget spent) and so the value will be negative and
in red.

This commit reviews the calculation to really show the budget remaining.
Also, since the allocated budget could be a negative amount, the columns
have been renamed to be clearer as before.

The columns in the Budget section have been renamed as follow:
  Allocated => Forecast
  Spent => Actual
  Remaining => Difference

Task-2948441

closes odoo/enterprise#32728

X-original-commit: b001cfe602f24b91086a1f85a680ed4e779c4730
Signed-off-by: Laurent Stukkens (ltu) <ltu@odoo.com>