[FIX] spreadsheet{_edition}: global filter with empty field
Before this commit, it was not possible to save a filter without selecting
a field for each data source.
However, adding a data source after the definition of the filter (which
lead to an empty field for this data source) was allowed, which was
incoherent with the first point.

As it's ok to have an empty field, we allow this and warn the user that
the filter will not be applied to this data source.

This commit also fix an issue where a field could be an empty object (`{}`)
instead of `undefined`, which caused a traceback during the evaluation
of the filter.

Task-id 2959765

closes odoo/enterprise#30690

Signed-off-by: Lucas Lefèvre (lul) <lul@odoo.com>