[IMP] account: import soda files
SODA is a standard file format created by CODABOX to describe the Accounting Entries to create coming from the payment of salaries in Belgium. Users in Belgium can use CODABOX that serves as an aggregator of data coming from the various Social Secretariat (Partena, Securex, SD Worx, etc), compile them and produce them in that standard format.

The objective of this PR is to allow the user to import the SODA files they get from their CODABOX account and import them manually in Odoo to create their Miscelaneous entries.

When using Drag & Import or the Upload functionality on Journal Entries or via the Drag & Drop on a Miscellaneous Journal, detect if the file is a SODA and create the appropriate entry with all the lines included.

Dependent on l10n_be.

Task-2258457

closes odoo/enterprise#30680

Related: odoo/odoo#98904
Signed-off-by: Quentin De Paoli <qdp@odoo.com>