[FIX] spreadsheet: add bus module dependency
The spreadsheet module depends on bus, but don't mark it as a
dependency. As a result, single build of the spreadsheet module
fails, because the `ir.websocket` module, which is defined on
the bus module, can't be resolved. This commit fixes this issue
by adding the required dependency in the manifest.

closes odoo/enterprise#30688

Signed-off-by: Pierre Rousseau (pro) <pro@odoo.com>