[IMP] stock_barcode*: change buttons size
Makes the buttons smaller to gain space on small screen.
As little screens (e.g.: smartphone) are usually used in portrait
orientation, and bigger screens (e.g.: desktop, tablet) are usued in
landscape orientation, the buttons' size change is based on that instead
of the device's size.

task-2960938

X-original-commit: 0328a1e6fa0d954b3f0174631a9ab52186e40325
Part-of: odoo/enterprise#32481