[FIX] l10n_cl_edi: If there is no shared certificate in the company the cronjobs that send invoices and check invoices in the SII don't work because they don't count with a valid certificate to operate.
This occurs for a few installations where the users own a certificate for each of them.

After this commit:
add warning msgs for lack of shared certificate inviting to make the send and ask manually. Also advising that the receipt acknowledge has not been sent due to lack of this shared certificate

[FIX] The l10n_cl_is_there_shared_certificate field is defined in model res.partner, the fields defined in other models become related to this one.

[REF] Changed country_code by account_fiscal_country_id.code

[FIX] L10n staging: assigned True by default if company is not CL

closes odoo/enterprise#29460

Signed-off-by: William André (wan) <wan@odoo.com>