[FIX] account_disallowed_expenses_fleet: add vehicle_id as tax grouping key
Currently, the non-deductible part of a tax is not linked to
a vehicle, and is thus not grouped correctly in the disallowed
expenses report.

This PR adds vehicle_id as a tax grouping key and solves
this issue.

At the same time, 'vehicle_id' is added as an onchange
on invoice lines, to make sure it also triggers
the recomputation of taxes.

opw-2941494

closes odoo/enterprise#30800

X-original-commit: b181b0fc2c6289a2cd3b6670af2df9dd049b99b8
Signed-off-by: Olivier Colson (oco) <oco@odoo.com>
Signed-off-by: John Laterre (jol) <jol@odoo.com>