[IMP] account_accountant: link reconciled matching button
The matching button, which shows either "MATCH", "P", "Ax" or nothing
now directs to the list of the reconciled ("P" or "Ax") journal items
of that reconciliation. This is a temporary fix, as it should open
the reconciled items in the manual reconciliation view instead, but
something needs to be done for v16.

We take advantage of this change to convert it to OWL.

task-2959808

closes odoo/enterprise#30702

Related: odoo/odoo#98827
Signed-off-by: Laurent Smet <las@odoo.com>