[IMP] hr_payroll: include month and year in payslip pdf name
When downloading multiple payslips (of the same employee) it is
impossible to distinguish between them because the name of the pdf
only includes the name of the employee.

From now on, the month and year are also included in the pdf name to
be able to see at a glance which payslip the file contains.

task-2954981

closes odoo/enterprise#30744

Signed-off-by: Kevin Baptiste <kba@odoo.com>