[FIX] account_sepa: incorrect SEPA credit transfer in currency
Side effect of commit [1].

With this commit, the goal was to allow SEPA credit transfers for Swiss
companies instead of forcing the use of generic credit transfers.

However, other situations were encountered were some payments were
considered as SEPA while they should be generic instead. This commit
should solve the 2 issues identified:

1/ for Swiss companies, all warnings were deactivated. We should remove
that, as there is no real business need for this, and we now need to
properly detect SEPA and generic payments.

2/ for operations in currency other than EUR, we were not returning a
warning if the currency of the company was not EUR, which is often the
case for Swiss companies.

We should keep the existing intended behavior and not display useless
warning in case of Swiss companies, but we should do the filtering
outside of the function that computes the warnings to avoid side effects
as it is used elsewhere like in the function to compute the field
sct_generic.

[1]:https://github.com/odoo/enterprise/commit/f1b955609633f9c5debf36035425455a7fdb6ac3

closes odoo/enterprise#30770

closes odoo/enterprise#30812

X-original-commit: ac9cef53dc9a14fdd6480d20979f7819a51d2eaf
Signed-off-by: Olivier Colson (oco) <oco@odoo.com>
Signed-off-by: Alex Thuyls (alt) <alt@odoo.com>