[IMP] sign: improve signature wizard layout
Several options have been added in the Sign flow which made the
wizard kind of clogged up.

This commit slightly cleans the view by moving several fields to
an 'options' tab, so that users can directly concentrate on the
data that matters the most: who will sign, and the message
attached to the signature request.

Task-2964300

closes odoo/enterprise#30793

Signed-off-by: Arnaud Joset <arj@odoo.com>