[FIX] sign: skip test if no credit field
Issue:
The tests are supposed to pass when the module sign is in standalone
and they fail here because the field credit is only added with the
account module. No monetary field is set on res.partner with only
the sign module that's why we use credit from account.

Solution:
The test is run only if the field credit exists (if the account module
is installed)

opw-2920379

closes odoo/enterprise#31028

X-original-commit: 0bdd10be8d75e5193851cd4fefda4cb877fc931b
Signed-off-by: Fockedey Martin (mafo) <mafo@odoo.com>
Signed-off-by: Xavier Dollé (xdo) <xdo@odoo.com>