[FIX] hr_payroll: cleanup todolist properly
The html field is not yet converted to owl.
This causes some issues with how we have to display the notes in the
payroll dashboard.
Since we have to pass through ComponentAdapter, the code assumes that
the widget properly cleans up between runs of `_reset` and `_render`
however some parts of the widget are never cleaned up, for instance the
`QWebPlugin`.
This commit makes sure that we remove it if it exists before re
rendering.

closes odoo/enterprise#30894

Signed-off-by: William Braeckman (wbr) <wbr@odoo.com>