[IMP] sign,sign_itsme: allow signing without IAP credits
When an instance doesn't have IAP credits for SMS (or now for itsme) but still
uses an SMS/itsme-enabled role in a signature request, the signatory is locked.
This is overall a bad experience. This commit improves this by making the lack
of credit invisible to the signatory: if they're supposed to identify by SMS
or itsme but the database doesn't have any credit anymore, the signature
proceeds as normal but we warn the user who requested the signature of the
issue, so that they may act upon it: either they decide that the extra
authentication was not really necessary, or they think it is and they can
request the signature again after buying credits.

We also want to warn users about low credits in advance so they can avoid the
situation in the first place.

task-2963665

closes odoo/enterprise#30844

Signed-off-by: Arnaud Joset <arj@odoo.com>