[FIX] account_asset: error with assets having multiple related expenses
Side effect of recent commit [1].

An asset can have original_move_line_ids linked to several moves. This
was not handled properly in the mentioned commit as a singleton error
was raised. We should adapt it by mapping field to 'debit' in case
any of the moves is of type 'in_refund' instead.

[1]:https://github.com/odoo/enterprise/commit/10c7711412d2a0b7eedc3eea512ed1424ed031eb

closes odoo/enterprise#30944

X-original-commit: ef4ff177d61e7743553691f2ba4b903fc33084af
Signed-off-by: Quentin De Paoli <qdp@odoo.com>
Signed-off-by: Alex Thuyls (alt) <alt@odoo.com>