[FIX] account_asset: error when opening depreciation schedule report
Side effect of recent commit [1].

An asset can have original_move_line_ids linked to several moves. This
was not handled properly in the mentioned commit as a singleton error
was raised. We should adapt it by mapping is_negative_asset to 'debit'
in case any of the moves is of type 'in_refund' instead.

[1]:https://github.com/odoo/enterprise/commit/99f84c4991e0bdd4f4a8073fedf715304c456496

closes odoo/enterprise#30997

X-original-commit: 36d46e29ade24d27d84b503947a564533fc0c36a
Signed-off-by: Florian Gilbert (flg) <flg@odoo.com>
Signed-off-by: Alex Thuyls (alt) <alt@odoo.com>