[FIX] account_saft: Missing currency_code
Before this commit:
Trying to generate a SAF-T report from the General Ledger (Norway) produces a traceback when there are account_move_lines in another currency than the company main currency.

After this commit:
Lines of the SAF-T report with a different currency than the company main currency are correctly generated.

OPW-2959672

closes odoo/enterprise#30923

X-original-commit: 6dc80e5fc2de18c05bc2ea56f78e0da8a787b35a
Signed-off-by: Laurent Smet <las@odoo.com>
Signed-off-by: thco-odoo <thco@odoo.com>