[FIX] timesheet_grid_holidays: holiday timesheet can't be deleted
When creating time off timesheet from the timesheet app,the newly
created timesheet is connected to a leave request while it shouldn't
be linked to any leave request and in consequence you can only delete
this new timesheet by deleting the leave request.

Explanation:
To create the new timesheet timesheet_grid makes a copy of another
timesheet on the same line and that makes that the holiday_id (used
to link timesheet and leave request) is also copied.

Solution:
When copying a timesheet the holiday_id is not copied

This commit add the test corresponding to the fix

opw-2951062

closes odoo/enterprise#31347

X-original-commit: 400c8c809180adb28c42bedbb8df16c9a7380551
Related: odoo/odoo#100211
Signed-off-by: Kevin Baptiste <kba@odoo.com>