[FIX] l10n_mx_reports: no reversed/reversal entry on diot report
With this commit we avoid normal entries that are reversed and their
reversal entries to appear on diot report. Only entries coming
from credit notes of vendor bills are taken into account.

Reversed entries have a reversal_move_id and reversals, a
reversed_entry_id. Entries from vendor bills and the ones
from credit notes have not. Then, we filter the two fields
on the domain which will be passed for the query's where clause.

Explanation: Those entries are the cash basis entries generated for taxes
on payment.  When you unreconcile a payment, a reverse cash basis entry
is generated because the original one needs to be undone (as it can
not be considered as paid anymore).  Those are the only cash based
entries which have the reverse normally, so we can assume that these
are the cases for which you undid (unreconciled) something and the
DIOT does not need to include those.

opw-2950903

closes odoo/enterprise#32586

X-original-commit: a231c306edb5e2dcb7afac760c32e24eb2790b01
Signed-off-by: Josse Colpaert <jco@odoo.com>
Signed-off-by: Guillaume Vanleynseele (guva) <guva@odoo.com>