[FIX] project_enterprise: display next recurring dates
Bug:
ProjectTaskRecurrence._get_next_recurring_dates() has a param start_date,
which is the return value of Task._get_recurrence_start_date().
If the recurrence is in months, start_date is compared with
repeat_until, which is a date.
As Task.planned_date_begin is a datetime, it crashes.

Fix:
Return planned_date_begin.date()

closes odoo/enterprise#30932

X-original-commit: df7f09e3b08720a6f698bfd4a7b7553915fdb7ca
Signed-off-by: Laurent Stukkens (ltu) <ltu@odoo.com>