[IMP] knowledge: change duplicate article to copy article
Purpose
=======

Allow users to create their own copies of restricted articles.

Specifications
==============

Rename "duplicate" action in the kebab menu to "Create a copy". This action
creates a copy of an article in the user's Private section. It does not copy
the child articles, members,... Only the name, body, cover and icon.

Task-2919746 (Knowledge: improve copy / duplicate)

Part-of: odoo/enterprise#30908