[IMP] knowledge: add archived banner on archived articles
Before this commit, there was no way to know, in the form view, that an article
was archived.

After this commit, a warning banner is displayed at the top of the article to
inform the users that they are reading an archived article. From taht banner,
they can directly unarchive the article.

Task-2850521 (Knowledge: trash management)

Part-of: odoo/enterprise#30908