[FIX] account_report : ellipsis on reports
The problem was that when a text was too long on a table cell, the cell will increase to fit the text in a single line. This made some reports (depreciation schedule for example) hard to read because some columns will be extremely long.

The way it works now is that the text will take all the space available in the column without making it grow unless the column is smaller than 320px. In that case, the column will grow until 320px.

We also allow the first column to grow until the next column with text to optimize space. We take into account the level of the line and if the line has children or is a children. That helps keep the report coherent and easy to read.

Task-id : 2965841

closes odoo/enterprise#31014

X-original-commit: 0c830e9274627d4132c410b22611a780d5a33247
Signed-off-by: Ruben Gomes <rugo@odoo.com>
Signed-off-by: Olivier Colson (oco) <oco@odoo.com>