[FIX] web_gantt: can't modify menu with studio
Steps to reproduce:
- open studio on planning module base gantt view
- modify the menu organization

Current behavior:
Traceback : TypeError: Cannot read properties
of undefined (reading 'dataset')

Expected behavior:
No traceback

Explanation:
When opening the studio edit menu in a gantt view,
the drop function of the draggable elements in the
edit menu form is set by the gantt view.

Solution:
We restrict the set of the drop function in the
grantt view to grantt elements.

opw-2964319

closes odoo/enterprise#31184

X-original-commit: 16c0aefcb57d2645111b8ddd47de88872fdc7a53
Signed-off-by: Aaron Bohy (aab) <aab@odoo.com>