[IMP] sale_subscription: improve performance of SO logs domains
After 8fe6c96558bc9fb9f3655fa52fcbea3fea38afdd, subscriptions have been
merged into the sale_order table, which can make the table cardinality
much bigger than it used to be.

The former sale.subscription.log model is also re-created as sale.order.log,
and is now attached to sales.order rows rather than a subscription.

However, in a service company, the number of sales orders may be much
larger than the number of actual subscriptions. At the very least, it
will accumulate more "transactional" SOs to create/renew/upsell the
subscriptions.

The ORM does not know about this, and whenever a domain term refers to
the `order_id` of SO logs, it will evaluate term on the full sale_order
table, not just the subset of SOs that are subscriptions.

Technically, a domain like this:

     ```[('order_id.stage_category', '=', 'progress')]```

would be translated into something like the following SQL:

    ```SELECT id FROM sale_order_log WHERE order_id IN
          (SELECT id FROM sale_order where stage_category = 'progress')```

This can be sub-optimal when the overall number of sale_order rows
is much larger than the number of actual subscriptions, i.e.
the number of sale_order rows that may actually have sale_order_log
lines.

Because of this, the SQL execution times can be much slower than the
previous versions using the former sale.subscription.log model.

Activating `auto_join` on the `order_id` relationship is safe
permission-wise, because the same ACLs apply to SO logs and to the
corresponding SOs. It allows the ORM to build a JOIN query that
the database can optimize, if the left side has a much lower cardinality
than the right side. For example, it could use a single Nested Loop
strategy with PK Index Scans on the right side, rather than a Merge Join
based on a costly full Seq Scan of the right side (or even multiple
such scans if there are several domain terms using `order_id`).

closes odoo/enterprise#30941

X-original-commit: dbafb1dce7d2ee33483bf0ea497a23143c36aba1
Signed-off-by: Olivier Dony <odo@odoo.com>
Signed-off-by: Arnaud Joset <arj@odoo.com>