[FIX] knowledge: fix missing record error when removing members
When the user removes a member from the share panel, the user will get
a missing record error. The error occurs in the `article_remove_member`
method: In that method, we fetch the partner from the member after
deleting it. When we try to fetch the partner, the model raises an error
as the member partner no longer exists.

To fix the issue, we will backup a reference of the partner associated
to the member we want to delete before deleting it. We will then use that
reference in the following instructions.

Task-2968657 (Knowledge: fix error when removing members)

Part-of: odoo/enterprise#30908