[IMP] helpdesk_timesheet: compute total time wo checking security rules
Before this commit total time on timesheet for helpdesk team
computed according to user's access to timesheet and security
rules on account.analytic.line are quite complex.

So, in this commit compute those fields with sudo to ease the
perfomance of helpdesk app.

task-2890184

closes odoo/enterprise#30950

Related: odoo/upgrade#3843
Related: odoo/odoo#99372
Signed-off-by: Xavier Bol (xbo) <xbo@odoo.com>