[FIX] knowledge: remove unreachable article link from share panel
In Knowledge, an article can inherit its permissions and its members from
one of its parents. When it happens, the share panel includes a link to
the article giving new permissions or members.

When the user does not have access to the article, the user will get an
access error by clicking on the provided link. To avoid that, we will no
longer display a link in the share panel if the user does not have access
to the target article.

Task-2968511 (Knowledge: remove unreachable links from share panel)

Part-of: odoo/enterprise#30908