[FIX] project_timesheet_forecast: remove the empty row in all timesheet menu
Issue:
Suppose employee A does not have a planning-based timesheet, but project and task based
empty rows appear in all timesheets. Because its searches according to the current user's employee

In this commit
project and task based empty rows appear in my timesheet only not All timesheets. Because
the specification of the task-2859663

Reproduction steps:

  -Install project_timesheet_forecast
  -Open Planning App
  -Create a planning slot
  -Generate the timesheet
  -Open the timesheet App > All timesheet
  -Search the timesheet of any employee

task-2955983

closes odoo/enterprise#33131

X-original-commit: 6c8d400eb040dc246a164547b93ad4d49af9a1cd
Signed-off-by: Xavier <xbo@odoo.com>