[FIX] l10n_be_intrastat: make export function public
Currently, using the xml export button on the Belgian results in an
access error because the function it is attempting to call is a private
method.

The fix removes the underscore from the begining of the method and
updates the name of the function in the place where it is called.

closes odoo/enterprise#30977

Signed-off-by: Josse Colpaert <jco@odoo.com>