[FIX] documents: fix traceback in search panel
The API of the orm_service got changed during the development of the
documents B2B, unfortunately it made one of the function cause a
traceback.
This was uncaught since the call is mocked during the tests and does not
do the exact same error checking as the python server.

closes odoo/enterprise#30971

Signed-off-by: William Braeckman (wbr) <wbr@odoo.com>