[FIX] currency_rate_live: update code new website
Steps to reproduce:
- Have a company with main currency as AED and multicurrency activated
- In setting, select as a service UAE Central Bank, interval manually
- Run the Update button

Issue:
- no data recovered

Cause:
The website has been updated and the URL has been changed. And the website uses JS-generated content.

Solution:
Use the request used by the websiute to retrieve the data. The problem is that the table as arabic entries.
To solve this issue, I added the arabic entries in the MAP_CURRENCIES dictionnary.

opw-2964066

closes odoo/enterprise#31131

X-original-commit: b6190313f11bd8be6d14de647ed64da3b5ab3ab3
Signed-off-by: William André (wan) <wan@odoo.com>