[IMP] knowledge: change the default value of the article title
Currently, in the knowledge module, the default value for the article title
is "New Article". A new article does not have a default title, so it is
pointless to use the default title "New Article" when creating an article.

This commit changes the default value of the article title from
"New Article" to "Untitled" and adapts the test tours accordingly.

task-2907572

closes odoo/enterprise#31004

Signed-off-by: David Beguin (dbe) <dbe@odoo.com>