[ADD] mrp_subcontracting_studio: Avoid incompatibilities
There are imcompatibilites between studio and the webclient of
mrp_subcontracting :
To use the portal, the home_menu needs to be removed.
The problem is, some services linked to studio need home_menu to work.
So we disabled them in the mrp_subcontracting_enterprise module before.

Problem is, now ListRenderer is patched in web_studio to add the use of
the 'studio' service, the same that was disabled earlier.

Adding this bridge module can remove the uneccessary files from studio
leading into the subcontracting webclient.

Task-2885757

Part-of: odoo/enterprise#31024