[IMP] helpdesk: use standard m2m for helpdesk team followers
Replaces the empty chatter in the helpdesk team view by a m2m_tags field
to allow managing followers of the team.

The field is on the right so the view is balanced by default. It shifts
the auto_assignment field down to be next to 'visibility' and so that
checking auto_assignment does not move any field down.

Make "new ticket" subtype default so people added as team followers
receive notifications for new tickets.

Task-2804888

closes odoo/enterprise#29284

Related: odoo/odoo#95645
Signed-off-by: Thibault Delavallee (tde) <tde@openerp.com>