[IMP] knowledge: Improve portal virality
Purpose
=======
Encourages portal users that find Knowledge pretty cool
to use it themselves in their own companies.

Specifications
==============
Whether portal users are logged in or not, if they are
interested in Knowledge, this commits provides a front-end
way (button or link) to redirect them on the Odoo trial form
with the Knowledge app already selected.

UTMs are added to the link to keep track of how many leads
it generates.

Task-2936725

closes odoo/enterprise#30578

Signed-off-by: Thibault Delavallee (tde) <tde@openerp.com>